// engineering/frontend/types/job.ts
export interface GeoCoordinatesBlock {
  id: number;
  lat: number;
  lng: number;
  province: string;
}

// --- Vacancy Meta Block ---
export interface VacancyMetaBlock {
  posted_at: string;
  source: string;
  standardized_category: string;
  seniority: string;
  geo: GeoCoordinatesBlock;
  confidence_score: number;
  ai_version: string;
  error: boolean;
}
// ...
export interface JobVacancy {
  id: string;
  employer: string;
  job_role: string;
  job_type: string;
  key_responsibilities: string;
  qualifications: string;
  location: string;
  offers: string;
  is_remote: boolean;
  skills: string[];
  meta_data: VacancyMetaBlock;
}